Add updater plugin
