+++ src-tauri/Cargo.toml
@@ -25,3 +25,6 @@ tauri-plugin-dialog = "2.0.0-beta.2"
 tauri-plugin-shell = "2.0.0-beta.2"
 thiserror = "1.0.57"
 tokio = { version = "1.36.0", features = ["macros", "time"] }
+
+[target."cfg(not(any(target_os = \"android\", target_os = \"ios\")))".dependencies]
+tauri-plugin-updater = "2.0.0-beta.2"
