+++ src-tauri/capabilities/dialog.json
@@ -1,15 +1,15 @@
 {
-    "$schema": "../gen/schemas/desktop-schema.json",
-    "identifier": "dialog-capability",
-    "windows": [
-        "main"
-    ],
-    "platforms": [
-        "linux",
-        "macOS",
-        "windows"
-    ],
-    "permissions": [
-        "dialog:allow-confirm"
-    ]
+  "$schema": "../gen/schemas/desktop-schema.json",
+  "identifier": "dialog-capability",
+  "windows": [
+    "main"
+  ],
+  "platforms": [
+    "linux",
+    "macOS",
+    "windows"
+  ],
+  "permissions": [
+    "dialog:allow-confirm"
+  ]
 }
