+++ src-tauri/capabilities/event.json
@@ -1,16 +1,16 @@
 {
-    "$schema": "../gen/schemas/desktop-schema.json",
-    "identifier": "event-capability",
-    "windows": [
-        "main"
-    ],
-    "platforms": [
-        "linux",
-        "macOS",
-        "windows"
-    ],
-    "permissions": [
-        "event:allow-listen",
-        "event:allow-unlisten"
-    ]
+  "$schema": "../gen/schemas/desktop-schema.json",
+  "identifier": "event-capability",
+  "windows": [
+    "main"
+  ],
+  "platforms": [
+    "linux",
+    "macOS",
+    "windows"
+  ],
+  "permissions": [
+    "event:allow-listen",
+    "event:allow-unlisten"
+  ]
 }
