+++ src-tauri/capabilities/updater.json
@@ -1,6 +1,6 @@
 {
   "$schema": "../gen/schemas/desktop-schema.json",
-  "identifier": "shell-capability",
+  "identifier": "updater-capability",
   "windows": [
     "main"
   ],
@@ -10,6 +10,6 @@
     "windows"
   ],
   "permissions": [
-    "shell:allow-open"
+    "updater:allow-check"
   ]
 }
