+++ src-tauri/capabilities/webview.json
@@ -1,15 +1,15 @@
 {
-    "$schema": "../gen/schemas/desktop-schema.json",
-    "identifier": "webview-capability",
-    "windows": [
-        "main"
-    ],
-    "platforms": [
-        "linux",
-        "macOS",
-        "windows"
-    ],
-    "permissions": [
-        "webview:allow-internal-toggle-devtools"
-    ]
+  "$schema": "../gen/schemas/desktop-schema.json",
+  "identifier": "webview-capability",
+  "windows": [
+    "main"
+  ],
+  "platforms": [
+    "linux",
+    "macOS",
+    "windows"
+  ],
+  "permissions": [
+    "webview:allow-internal-toggle-devtools"
+  ]
 }
