+++ src-tauri/gen/schemas/capabilities.json
@@ -1,1 +1,1 @@
-{"dialog-capability":{"identifier":"dialog-capability","description":"","local":true,"windows":["main"],"permissions":["dialog:allow-confirm"],"platforms":["linux","macOS","windows"]},"event-capability":{"identifier":"event-capability","description":"","local":true,"windows":["main"],"permissions":["event:allow-listen","event:allow-unlisten"],"platforms":["linux","macOS","windows"]},"shell-capability":{"identifier":"shell-capability","description":"","local":true,"windows":["main"],"permissions":["shell:allow-open"],"platforms":["linux","macOS","windows"]},"webview-capability":{"identifier":"webview-capability","description":"","local":true,"windows":["main"],"permissions":["webview:allow-internal-toggle-devtools"],"platforms":["linux","macOS","windows"]}}
+{"dialog-capability":{"identifier":"dialog-capability","description":"","local":true,"windows":["main"],"permissions":["dialog:allow-confirm"],"platforms":["linux","macOS","windows"]},"event-capability":{"identifier":"event-capability","description":"","local":true,"windows":["main"],"permissions":["event:allow-listen","event:allow-unlisten"],"platforms":["linux","macOS","windows"]},"shell-capability":{"identifier":"shell-capability","description":"","local":true,"windows":["main"],"permissions":["shell:allow-open"],"platforms":["linux","macOS","windows"]},"updater-capability":{"identifier":"updater-capability","description":"","local":true,"windows":["main"],"permissions":["updater:allow-check"],"platforms":["linux","macOS","windows"]},"webview-capability":{"identifier":"webview-capability","description":"","local":true,"windows":["main"],"permissions":["webview:allow-internal-toggle-devtools"],"platforms":["linux","macOS","windows"]}}
