+++ src-tauri/src/lib.rs
@@ -39,6 +39,7 @@ pub fn run() {
         .setup(setup)
         .plugin(tauri_plugin_dialog::init())
         .plugin(tauri_plugin_shell::init())
+        .plugin(tauri_plugin_updater::Builder::new().build())
         .invoke_handler(tauri::generate_handler![
             commands::login,
             commands::logout,
