+++ src-tauri/tauri.conf.json
@@ -32,5 +32,14 @@
       "icons/icon.icns",
       "icons/icon.ico"
     ]
+  },
+  "plugins": {
+    "updater": {
+      "active": true,
+      "endpoints": [
+        "https://sugyan.com/tarium/update.json"
+      ],
+      "pubkey": "dW50cnVzdGVkIGNvbW1lbnQ6IG1pbmlzaWduIHB1YmxpYyBrZXk6IDYzNjEyMkE5NUVGMTM0RjEKUldUeE5QRmVxU0poWTdQbFE1cnFiRVh0RzVVNzQ3V2dwRms3SHRsVXN6TkcvWUtRRlBOWWJoTUEK"
+    }
   }
 }
