+++ src/main.tsx
@@ -1,34 +1,10 @@
 import { StrictMode } from "react";
 import ReactDOM from "react-dom/client";
-import { RouterProvider, createBrowserRouter } from "react-router-dom";
+import App from "./App";
 import "./index.css";
-import Home from "./routes/home";
-import Root from "./routes/root";
-import Signin from "./routes/signin";
-
-const router = createBrowserRouter([
-  {
-    path: "/",
-    element: <Root />,
-    children: [
-      {
-        path: "/home",
-        element: <Home />,
-      },
-    ],
-  },
-  {
-    path: "/signin",
-    element: <Signin />,
-  },
-]);
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <StrictMode>
-    <div className="dark">
-      <div className="dark:text-gray-200 dark:bg-gray-800">
-        <RouterProvider router={router} />
-      </div>
-    </div>
+    <App />
   </StrictMode>
 );
